Extract angular releated functions
